# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FinSight - Personal financial tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  name: string;
  /** Used to convert expenses into hours of life spent */
  hourlyWage: number;
  currency: string;
  /** @nullable */
  monthlyIncomeGoal?: number | null;
  /** @nullable */
  monthlySavingsGoal?: number | null;
  createdAt: string;
}

export interface ProfileUpdate {
  name?: string;
  hourlyWage?: number;
  currency?: string;
  /** @nullable */
  monthlyIncomeGoal?: number | null;
  /** @nullable */
  monthlySavingsGoal?: number | null;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface Transaction {
  id: number;
  amount: number;
  type: TransactionType;
  category: string;
  description: string;
  date: string;
  /**
     * 1 (no regret) to 5 (maximum regret) — user-rated after purchase
     * @nullable
     */
  regretScore?: number | null;
  /**
     * Mood tag linked at time of transaction
     * @nullable
     */
  moodAtPurchase?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface TransactionInput {
  /** @minimum 0.01 */
  amount: number;
  type: TransactionInputType;
  category: string;
  description: string;
  date: string;
  moodAtPurchase?: string;
  notes?: string;
}

export type TransactionUpdateType = typeof TransactionUpdateType[keyof typeof TransactionUpdateType];


export const TransactionUpdateType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface TransactionUpdate {
  amount?: number;
  type?: TransactionUpdateType;
  category?: string;
  description?: string;
  date?: string;
  moodAtPurchase?: string;
  notes?: string;
}

export interface RegretRating {
  /**
     * @minimum 1
     * @maximum 5
     */
  score: number;
  notes?: string;
}

export interface MoodLog {
  id: number;
  /** One of: happy, calm, stressed, anxious, bored, sad, excited, angry */
  mood: string;
  /** @nullable */
  note: string | null;
  /**
     * Auto-computed spending on this date
     * @nullable
     */
  spendingAmountToday?: number | null;
  loggedAt: string;
}

export interface MoodLogInput {
  mood: string;
  note?: string;
}

export interface ProjectionPoint {
  month: number;
  value: number;
}

export interface Scenario {
  id: number;
  name: string;
  /** Human-readable description of what changes (e.g. "Cancel Netflix + cook at home 4x/week") */
  changeDescription: string;
  /** How much more (positive) or less (negative) would be saved per month */
  monthlySavingsDelta: number;
  /** Number of years to project into the future */
  projectionYears: number;
  /** Assumed annual investment return rate (e.g. 0.07 for 7%) */
  annualReturnRate?: number;
  /** Total compounded savings at end of projection period */
  projectedSavings: number;
  monthlyBreakdown?: ProjectionPoint[];
  createdAt: string;
}

export interface ScenarioInput {
  name: string;
  changeDescription: string;
  monthlySavingsDelta: number;
  /**
     * @minimum 1
     * @maximum 40
     */
  projectionYears: number;
  annualReturnRate?: number;
}

export interface CategoryStat {
  category: string;
  total: number;
  count: number;
  percentage: number;
}

export interface DashboardSummary {
  totalIncome: number;
  totalExpenses: number;
  netBalance: number;
  topCategory: string;
  transactionCount: number;
  avgDailySpend: number;
  /** Percentage of income saved (0-100) */
  savingsRate: number;
  currentMonthExpenses: number;
  currentMonthIncome: number;
  categoryBreakdown?: CategoryStat[];
  recentTransactions?: Transaction[];
}

export interface SpendingDna {
  dominantCategory: string;
  /** e.g. 'The Comfort Seeker', 'The Impulse Artist', 'The Disciplined Saver' */
  spendingPersonality: string;
  /** Day of week with highest average spend */
  peakSpendingDay: string;
  /** 0-100 score indicating impulsive spending tendency */
  impulseScore: number;
  /** 0-100 score for how consistent spending patterns are */
  consistencyScore: number;
  /** 0-100 score for how spread across categories spending is */
  diversityScore: number;
  /** Up to 5 detected behavioral patterns in plain language */
  topPatterns: string[];
}

export interface TimeCostItem {
  transactionId: number;
  description: string;
  amount: number;
  /** amount / hourlyWage */
  hoursOfWork: number;
  category: string;
  date: string;
}

export interface MoodCorrelationPoint {
  mood: string;
  avgSpend: number;
  count: number;
  totalSpend: number;
}

export interface CategoryRegretStat {
  category: string;
  avgRegretScore: number;
  count: number;
}

export interface RegretAnalysis {
  avgRegretScore: number;
  mostRegrettedCategory: string;
  leastRegrettedCategory: string;
  /** Total money spent on transactions with regret score >= 4 */
  totalRegretted: number;
  regretByCategory: CategoryRegretStat[];
  topRegrettedTransactions: Transaction[];
}

export type ListTransactionsParams = {
category?: string;
type?: ListTransactionsType;
limit?: number;
};

export type ListTransactionsType = typeof ListTransactionsType[keyof typeof ListTransactionsType];


export const ListTransactionsType = {
  income: 'income',
  expense: 'expense',
} as const;